+++ bailiancontrol-20240816/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-06-23 Version: 1.0.2
+- Update API ListApiKeys: add request parameters order.
+- Update API ListApiKeys: add request parameters orderBy.
+
+
 2026-03-30 Version: 1.0.1
 - Update API ListApiKeys: add response parameters Body.apiKeys.$.authSetModel.